Task
Task: From the integer array X(N), write to the array Y(k) all elements exceeding the arithmetic mean of all elements of the array X. Delete all elements in arrays that are not prime numbers. In the program, write functions: forming an array of Y, determining the arithmetic mean of elements, deleting an element.